requirements.txt
flask
flask-restx
requests
pytest
pytest-html
pyhamcrest
jsonschema
